move stuff to helper class
